Collision
Was getting pissed with all the math I had to do to make the sprites work how I wanted, so I just changed the sprites.

Added object collision/interaction. I wanna come up with more of a system for the interaction probably gunna make a function, but this is good for now.